Split readwrite-splitting ANTLR g4 file and visitor
