+++ shardingsphere-distsql-parser/shardingsphere-distsql-parser-engine/src/test/java/org/apache/shardingsphere/distsql/parser/api/DistSQLStatementParserEngineTest.java
@@ -21,21 +21,16 @@ import org.apache.shardingsphere.distsql.parser.segment.DataSourceSegment;
 import org.apache.shardingsphere.distsql.parser.segment.rdl.DatabaseDiscoveryRuleSegment;
 import org.apache.shardingsphere.distsql.parser.segment.rdl.EncryptColumnSegment;
 import org.apache.shardingsphere.distsql.parser.segment.rdl.EncryptRuleSegment;
-import org.apache.shardingsphere.distsql.parser.segment.rdl.ReadwriteSplittingRuleSegment;
 import org.apache.shardingsphere.distsql.parser.statement.rdl.alter.impl.AlterDatabaseDiscoveryRuleStatement;
 import org.apache.shardingsphere.distsql.parser.statement.rdl.alter.impl.AlterEncryptRuleStatement;
-import org.apache.shardingsphere.distsql.parser.statement.rdl.alter.impl.AlterReadwriteSplittingRuleStatement;
 import org.apache.shardingsphere.distsql.parser.statement.rdl.create.AddResourceStatement;
 import org.apache.shardingsphere.distsql.parser.statement.rdl.create.impl.CreateDatabaseDiscoveryRuleStatement;
 import org.apache.shardingsphere.distsql.parser.statement.rdl.create.impl.CreateEncryptRuleStatement;
-import org.apache.shardingsphere.distsql.parser.statement.rdl.create.impl.CreateReadwriteSplittingRuleStatement;
 import org.apache.shardingsphere.distsql.parser.statement.rdl.drop.DropResourceStatement;
 import org.apache.shardingsphere.distsql.parser.statement.rdl.drop.impl.DropDatabaseDiscoveryRuleStatement;
 import org.apache.shardingsphere.distsql.parser.statement.rdl.drop.impl.DropEncryptRuleStatement;
-import org.apache.shardingsphere.distsql.parser.statement.rdl.drop.impl.DropReadwriteSplittingRuleStatement;
 import org.apache.shardingsphere.distsql.parser.statement.rql.show.impl.ShowDatabaseDiscoveryRulesStatement;
 import org.apache.shardingsphere.distsql.parser.statement.rql.show.impl.ShowEncryptRulesStatement;
-import org.apache.shardingsphere.distsql.parser.statement.rql.show.impl.ShowReadwriteSplittingRulesStatement;
 import org.apache.shardingsphere.sql.parser.sql.common.statement.SQLStatement;
 import org.junit.Test;
 
@@ -59,28 +54,6 @@ public final class DistSQLStatementParserEngineTest {
             + "ds_1(HOST=127.0.0.1,PORT=3306,DB=test1,USER=ROOT,PASSWORD=123456);";
     
     private static final String RDL_DROP_RESOURCE = "DROP RESOURCE ds_0,ds_1";
-    
-    private static final String RDL_CREATE_STATIC_READWRITE_SPLITTING_RULE = "CREATE READWRITE_SPLITTING RULE ms_group_0 ("
-            + "WRITE_RESOURCE=primary_ds,"
-            + "READ_RESOURCES(replica_ds_0,replica_ds_1),"
-            + "TYPE(NAME=random)"
-            + ")";
-    
-    private static final String RDL_CREATE_DYNAMIC_READWRITE_SPLITTING_RULE = "CREATE READWRITE_SPLITTING RULE ms_group_1 ("
-            + "AUTO_AWARE_RESOURCE=group_0,"
-            + "TYPE(NAME=random,PROPERTIES(read_weight='2:1'))"
-            + ")";
-    
-    private static final String RDL_ALTER_READWRITE_SPLITTING_RULE = "ALTER READWRITE_SPLITTING RULE ms_group_0 ("
-            + "AUTO_AWARE_RESOURCE=group_0,"
-            + "TYPE(NAME=random,PROPERTIES(read_weight='2:1'))),"
-            + "ms_group_1 ("
-            + "WRITE_RESOURCE=primary_ds,"
-            + "READ_RESOURCES(replica_ds_0,replica_ds_1),"
-            + "TYPE(NAME=random)"
-            + ")";
-    
-    private static final String RDL_DROP_READWRITE_SPLITTING_RULE = "DROP READWRITE_SPLITTING RULE ms_group_0,ms_group_1";
     
     private static final String RDL_CREATE_DATABASE_DISCOVERY_RULE = "CREATE DB_DISCOVERY RULE ha_group_0 ("
             + "RESOURCES(resource0,resource1),"
@@ -113,8 +86,6 @@ public final class DistSQLStatementParserEngineTest {
             + "))";
     
     private static final String RDL_DROP_ENCRYPT_RULE = "DROP ENCRYPT RULE t_encrypt,t_encrypt_order";
-    
-    private static final String RQL_SHOW_READWRITE_SPLITTING_RULES = "SHOW READWRITE_SPLITTING RULES FROM readwrite_splitting_db";
     
     private static final String RQL_SHOW_DB_DISCOVERY_RULES = "SHOW DB_DISCOVERY RULES FROM db_discovery_db";
     
@@ -179,67 +150,6 @@ public final class DistSQLStatementParserEngineTest {
         assertTrue(sqlStatement instanceof DropResourceStatement);
         assertThat(((DropResourceStatement) sqlStatement).getNames().size(), is(2));
         assertTrue(((DropResourceStatement) sqlStatement).getNames().containsAll(Arrays.asList("ds_0", "ds_1")));
-    }
-    
-    @Test
-    public void assertParseStaticReadwriteSplittingRule() {
-        SQLStatement sqlStatement = engine.parse(RDL_CREATE_STATIC_READWRITE_SPLITTING_RULE);
-        assertTrue(sqlStatement instanceof CreateReadwriteSplittingRuleStatement);
-        CreateReadwriteSplittingRuleStatement statement = (CreateReadwriteSplittingRuleStatement) sqlStatement;
-        assertThat(statement.getRules().size(), is(1));
-        List<ReadwriteSplittingRuleSegment> readwriteSplittingRuleSegments
-                = new ArrayList<>(((CreateReadwriteSplittingRuleStatement) sqlStatement).getRules());
-        assertThat(readwriteSplittingRuleSegments.get(0).getName(), is("ms_group_0"));
-        assertThat(readwriteSplittingRuleSegments.get(0).getWriteDataSource(), is("primary_ds"));
-        assertThat(readwriteSplittingRuleSegments.get(0).getReadDataSources(), is(Arrays.asList("replica_ds_0", "replica_ds_1")));
-        assertThat(readwriteSplittingRuleSegments.get(0).getLoadBalancer(), is("random"));
-        assertThat(readwriteSplittingRuleSegments.get(0).getProps().size(), is(0));
-    }
-    
-    @Test
-    public void assertParseDynamicReadwriteSplittingRule() {
-        SQLStatement sqlStatement = engine.parse(RDL_CREATE_DYNAMIC_READWRITE_SPLITTING_RULE);
-        assertTrue(sqlStatement instanceof CreateReadwriteSplittingRuleStatement);
-        CreateReadwriteSplittingRuleStatement statement = (CreateReadwriteSplittingRuleStatement) sqlStatement;
-        assertThat(statement.getRules().size(), is(1));
-        List<ReadwriteSplittingRuleSegment> readwriteSplittingRuleSegments
-                = new ArrayList<>(((CreateReadwriteSplittingRuleStatement) sqlStatement).getRules());
-        assertThat(readwriteSplittingRuleSegments.get(0).getName(), is("ms_group_1"));
-        assertThat(readwriteSplittingRuleSegments.get(0).getAutoAwareResource(), is("group_0"));
-        assertNull(readwriteSplittingRuleSegments.get(0).getWriteDataSource());
-        assertNull(readwriteSplittingRuleSegments.get(0).getReadDataSources());
-        assertThat(readwriteSplittingRuleSegments.get(0).getLoadBalancer(), is("random"));
-        assertThat(readwriteSplittingRuleSegments.get(0).getProps().size(), is(1));
-        assertThat(readwriteSplittingRuleSegments.get(0).getProps().getProperty("read_weight"), is("'2:1'"));
-    }
-    
-    @Test
-    public void assertParseAlterReadwriteSplittingRule() {
-        SQLStatement sqlStatement = engine.parse(RDL_ALTER_READWRITE_SPLITTING_RULE);
-        assertTrue(sqlStatement instanceof AlterReadwriteSplittingRuleStatement);
-        AlterReadwriteSplittingRuleStatement statement = (AlterReadwriteSplittingRuleStatement) sqlStatement;
-        assertThat(statement.getRules().size(), is(2));
-        List<ReadwriteSplittingRuleSegment> readwriteSplittingRuleSegments
-                = new ArrayList<>(((AlterReadwriteSplittingRuleStatement) sqlStatement).getRules());
-        assertThat(readwriteSplittingRuleSegments.get(0).getName(), is("ms_group_0"));
-        assertThat(readwriteSplittingRuleSegments.get(0).getAutoAwareResource(), is("group_0"));
-        assertNull(readwriteSplittingRuleSegments.get(0).getWriteDataSource());
-        assertNull(readwriteSplittingRuleSegments.get(0).getReadDataSources());
-        assertThat(readwriteSplittingRuleSegments.get(0).getLoadBalancer(), is("random"));
-        assertThat(readwriteSplittingRuleSegments.get(0).getProps().size(), is(1));
-        assertThat(readwriteSplittingRuleSegments.get(0).getProps().getProperty("read_weight"), is("'2:1'"));
-        assertThat(readwriteSplittingRuleSegments.get(1).getName(), is("ms_group_1"));
-        assertThat(readwriteSplittingRuleSegments.get(1).getWriteDataSource(), is("primary_ds"));
-        assertThat(readwriteSplittingRuleSegments.get(1).getReadDataSources(), is(Arrays.asList("replica_ds_0", "replica_ds_1")));
-        assertThat(readwriteSplittingRuleSegments.get(1).getLoadBalancer(), is("random"));
-        assertThat(readwriteSplittingRuleSegments.get(1).getProps().size(), is(0));
-    }
-    
-    @Test
-    public void assertParseDropReadwriteSplittingRule() {
-        SQLStatement sqlStatement = engine.parse(RDL_DROP_READWRITE_SPLITTING_RULE);
-        assertTrue(sqlStatement instanceof DropReadwriteSplittingRuleStatement);
-        assertThat(((DropReadwriteSplittingRuleStatement) sqlStatement).getRuleNames(), is(Arrays.asList("ms_group_0", "ms_group_1")));
     }
     
     @Test
@@ -329,13 +239,6 @@ public final class DistSQLStatementParserEngineTest {
         SQLStatement sqlStatement = engine.parse(RDL_DROP_ENCRYPT_RULE);
         assertTrue(sqlStatement instanceof DropEncryptRuleStatement);
         assertThat(((DropEncryptRuleStatement) sqlStatement).getTables(), is(Arrays.asList("t_encrypt", "t_encrypt_order")));
-    }
-    
-    @Test
-    public void assertParseShowReadwriteSplittingRules() {
-        SQLStatement sqlStatement = engine.parse(RQL_SHOW_READWRITE_SPLITTING_RULES);
-        assertTrue(sqlStatement instanceof ShowReadwriteSplittingRulesStatement);
-        assertThat(((ShowReadwriteSplittingRulesStatement) sqlStatement).getSchema().get().getIdentifier().getValue(), is("readwrite_splitting_db"));
     }
     
     @Test
