+++ shardingsphere-features/shardingsphere-sharding/shardingsphere-sharding-distsql/src/main/resources/META-INF/services/org.apache.shardingsphere.distsql.parser.spi.RuleSQLStatementVisitorFacade
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.sharding.distsql.parser.ShardingRuleSQLStatementVisitorFacade
+org.apache.shardingsphere.sharding.distsql.parser.facade.ShardingRuleSQLStatementVisitorFacade
